requirements.txt
Pyside6==6.6.1
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
imageio==2.33.0